refactor: fix #22 & some stuff. See below
New `OTHER` collection in the database for different stuff.
Genshin Cookies are now stored in the db.
Fly Me To the Moon chapters are now stored in the db.
Rename `GlobalUserData` to `GlobalUser`.
Add properties for GlobalData model.
`GlobalData` are now stored users in the list.
Set `ssl_cert_reqs`. something wrong with my pc